Fixed Table of Contents Links and Color issues
Sections now have their numbers suppressed.
Previously, sections were defined with 'section*{}'.
The asterisk prevented the numbers from appearing in the TOC.
But the asterisk was also interfering with the hyperref package.
So sections are now definted without the asterisk.

somewhere in the LaTeX code, it is changing 'red' to 'RED', so I had to
create a special translation so that xcolor doesn't throw a fit.